+++ internal/app/types/entity_mongo.go
@@ -42,6 +42,9 @@ type Entity struct {
 
 func (entity *Entity) Validate() []error {
 	errs := []error{}
+
+	errs = append(errs, validateEmail(entity.Email)...)
+
 	if len(entity.EntityName) > 100 {
 		errs = append(errs, errors.New("Entity name length cannot exceed 100 characters."))
 	}
